requirements.txt
numpy
flask
huggingface_hub
requests
python-dateutil
flask-cors